move community_digest_data out of view into model to fix circular dependency issue
