+++ api/tests/test_learning_circle_api.py
@@ -3,6 +3,7 @@ from django.test import TestCase
 from django.test import Client
 from django.core import mail
 from django.utils import timezone
+from django.conf import settings
 
 from mock import patch
 
@@ -53,7 +54,7 @@ class TestLearningCircleApi(TestCase):
         lc = StudyGroup.objects.all().last()
         self.assertEqual(resp.json(), {
             "status": "created",
-            "url": "example.net/en/signup/75-harrington-{}/".format(lc.pk)
+            "url": "{}/en/signup/75-harrington-{}/".format(settings.DOMAIN, lc.pk)
         })
         self.assertEqual(StudyGroup.objects.all().count(), 5)
         self.assertEqual(lc.course.id, 3)
@@ -99,7 +100,7 @@ class TestLearningCircleApi(TestCase):
         lc = StudyGroup.objects.all().last()
         self.assertEqual(resp.json(), {
             "status": "created",
-            "url": "example.net/en/signup/75-harrington-{}/".format(lc.pk)
+            "url": "{}/en/signup/75-harrington-{}/".format(settings.DOMAIN, lc.pk)
         })
         self.assertEqual(StudyGroup.objects.all().count(), 5)
         self.assertEqual(lc.course.id, 3)
@@ -176,7 +177,7 @@ class TestLearningCircleApi(TestCase):
         lc = StudyGroup.objects.all().last()
         self.assertEqual(resp.json(), {
             "status": "created",
-            "url": "example.net/en/signup/75-harrington-{}/".format(lc.pk)
+            "url": "{}/en/signup/75-harrington-{}/".format(settings.DOMAIN, lc.pk)
         })
         self.assertEqual(StudyGroup.objects.all().count(), 5)
         data['start_date'] = '2018-01-01'
@@ -223,7 +224,7 @@ class TestLearningCircleApi(TestCase):
         lc = StudyGroup.objects.all().last()
         self.assertEqual(resp.json(), {
             "status": "created",
-            "url": "example.net/en/signup/75-harrington-{}/".format(lc.pk)
+            "url": "{}/en/signup/75-harrington-{}/".format(settings.DOMAIN, lc.pk)
         })
         self.assertEqual(StudyGroup.objects.all().count(), 5)
         self.assertEqual(lc.meeting_set.all().count(), 0)
@@ -277,7 +278,7 @@ class TestLearningCircleApi(TestCase):
         lc = StudyGroup.objects.all().last()
         self.assertEqual(resp.json(), {
             "status": "created",
-            "url": "example.net/en/signup/%D8%A7%D9%84%D8%B5%D8%AD%D8%A9-%D8%A7%D9%84%D9%86%D9%81%D8%B3%D9%8A%D8%A9-%D9%84%D9%84%D8%B7%D9%81%D9%84-{}/".format(lc.pk)
+            "url": "{}/en/signup/%D8%A7%D9%84%D8%B5%D8%AD%D8%A9-%D8%A7%D9%84%D9%86%D9%81%D8%B3%D9%8A%D8%A9-%D9%84%D9%84%D8%B7%D9%81%D9%84-{}/".format(settings.DOMAIN, lc.pk)
         })
         self.assertEqual(StudyGroup.objects.all().count(), 5)
