+++ learnwithpeople/settings.py
@@ -259,8 +259,8 @@ CELERYBEAT_SCHEDULE = {
     },
     'send_biweekly_community_digest': {
         'task': 'studygroups.tasks.send_biweekly_community_digest',
-        'schedule': crontab(day_of_month='1,15')
-    }
+        'schedule': crontab(day_of_month='1,15'),
+    },
 }
 
 LOGGING = {
