+++ studygroups/tests/test_tasks.py
@@ -4,6 +4,7 @@ from django.core import mail
 from django.contrib.auth.models import User
 from django.utils import timezone
 from django.utils.translation import get_language
+from django.conf import settings
 
 from mock import patch
 from freezegun import freeze_time
@@ -17,16 +18,8 @@ from studygroups.models import Rsvp
 from studygroups.models import Team
 from studygroups.models import TeamMembership
 from studygroups.models import accept_application
-from studygroups.tasks import generate_reminder
-from studygroups.tasks import send_reminder
 from studygroups.models import create_rsvp
 from studygroups.models import generate_all_meetings
-from studygroups.tasks import send_weekly_update
-from studygroups.tasks import send_learner_surveys
-from studygroups.tasks import send_facilitator_survey
-from studygroups.tasks import send_facilitator_survey_reminder
-from studygroups.tasks import send_final_learning_circle_report
-from studygroups.tasks import send_last_week_group_activity
 from studygroups.rsvp import gen_rsvp_querystring
 from studygroups.rsvp import check_rsvp_signature
 from studygroups.utils import gen_unsubscribe_querystring
@@ -80,7 +73,7 @@ class TestStudyGroupTasks(TestCase):
         sg.save()
         sg = StudyGroup.objects.get(pk=1)
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 0)
 
 
@@ -97,14 +90,14 @@ class TestStudyGroupTasks(TestCase):
         generate_all_meetings(sg)
         self.assertEqual(sg.meeting_set.active().count(), 6)
         self.assertTrue(sg.next_meeting().meeting_datetime() - now < datetime.timedelta(days=4))
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         #TODO check that email was sent to site admin
         #TODO test with unicode in generated email subject
 
         # Make sure we do it only once
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
 
     def test_no_reminders_for_old_studygroups(self):
@@ -120,7 +113,7 @@ class TestStudyGroupTasks(TestCase):
         sg.save()
         sg = StudyGroup.objects.get(pk=1)
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 0)
 
 
@@ -137,7 +130,7 @@ class TestStudyGroupTasks(TestCase):
         sg.save()
         sg = StudyGroup.objects.get(pk=1)
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 0)
 
     def test_generate_reminder_with_long_name_and_location(self):
@@ -155,14 +148,14 @@ class TestStudyGroupTasks(TestCase):
         generate_all_meetings(sg)
         self.assertEqual(sg.meeting_set.active().count(), 6)
         self.assertTrue(sg.next_meeting().meeting_datetime() - now < datetime.timedelta(days=4))
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         #TODO check that email was sent to site admin
         #TODO test with unicode in generated email subject
 
         # Make sure we do it only once
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
 
 
@@ -186,7 +179,7 @@ class TestStudyGroupTasks(TestCase):
 
         mail.outbox = []
         with freeze_time("2010-03-06 18:55:34"):
-            generate_reminder(sg)
+            tasks.generate_reminder(sg)
 
         self.assertEqual(len(mail.outbox), 1)
         self.assertEqual(Reminder.objects.all().count(), 1)
@@ -202,7 +195,7 @@ class TestStudyGroupTasks(TestCase):
 
         mail.outbox = []
         with freeze_time("2010-03-13 18:55:34"):
-            generate_reminder(sg)
+            tasks.generate_reminder(sg)
         self.assertEqual(len(mail.outbox), 1)
         self.assertEqual(Reminder.objects.all().count(), 2)
         self.assertEqual(Reminder.objects.filter(sent_at__isnull=True).count(), 1)
@@ -234,17 +227,17 @@ class TestStudyGroupTasks(TestCase):
         application.save()
         mail.outbox = []
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         self.assertEqual(len(mail.outbox), 1)
-        send_reminder(reminder)
+        tasks.send_reminder(reminder)
         self.assertEqual(len(mail.outbox), 3) # should be sent to facilitator & application
         self.assertEqual(mail.outbox[1].to[0], data['email'])
         self.assertFalse(send_message.called)
-        self.assertIn('https://example.net/{0}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={1}&attending=yes&sig='.format(get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
-        self.assertIn('https://example.net/{0}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={1}&attending=no&sig='.format(get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
-        self.assertIn('https://example.net/{0}/optout/confirm/?user='.format(get_language()), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={2}&attending=yes&sig='.format(settings.DOMAIN, get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={2}&attending=no&sig='.format(settings.DOMAIN, get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/optout/confirm/?user='.format(settings.DOMAIN, get_language()), mail.outbox[1].alternatives[0][0])
 
 
     @patch('studygroups.tasks.send_message')
@@ -265,17 +258,17 @@ class TestStudyGroupTasks(TestCase):
         application.save()
         mail.outbox = []
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         self.assertEqual(len(mail.outbox), 1)
-        send_reminder(reminder)
+        tasks.send_reminder(reminder)
         self.assertEqual(len(mail.outbox), 3) # should be sent to facilitator & application
         self.assertEqual(mail.outbox[1].to[0], data['email'])
         self.assertFalse(send_message.called)
-        self.assertIn('https://example.net/{0}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={1}&attending=yes&sig='.format(get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
-        self.assertIn('https://example.net/{0}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={1}&attending=no&sig='.format(get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
-        self.assertIn('https://example.net/{0}/optout/confirm/?user='.format(get_language()), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={2}&attending=yes&sig='.format(settings.DOMAIN, get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/rsvp/?user=test%40mail.com&study_group=1&meeting_date={2}&attending=no&sig='.format(settings.DOMAIN, get_language(), urllib.parse.quote(sg.next_meeting().meeting_datetime().isoformat())), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/optout/confirm/?user='.format(settings.DOMAIN, get_language()), mail.outbox[1].alternatives[0][0])
         self.assertIn('VEVENT', mail.outbox[1].attachments[0].get_payload())
 
 
@@ -308,7 +301,7 @@ class TestStudyGroupTasks(TestCase):
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         self.assertEqual(len(mail.outbox), 0)
-        send_reminder(reminder)
+        tasks.send_reminder(reminder)
         self.assertEqual(len(mail.outbox), 1)
         self.assertEqual(len(mail.outbox[0].bcc), 2)
         self.assertEqual(mail.outbox[0].bcc[0], data['email'])
@@ -333,19 +326,19 @@ class TestStudyGroupTasks(TestCase):
         application.save()
         mail.outbox = []
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
         self.assertEqual(len(mail.outbox), 1)
         mail.outbox = []
-        send_reminder(reminder)
+        tasks.send_reminder(reminder)
         self.assertEqual(len(mail.outbox), 2) # should be sent to facilitator & application
         self.assertEqual(mail.outbox[0].to[0], data['email'])
         self.assertEqual(mail.outbox[1].to[0], sg.facilitator.email)
         self.assertFalse(send_message.called)
-        self.assertNotIn('https://example.net/{0}/rsvp/'.format(get_language()), mail.outbox[1].alternatives[0][0])
-        self.assertIn('https://example.net/{0}/facilitator/'.format(get_language()), mail.outbox[1].alternatives[0][0])
-        self.assertNotIn('https://example.net/{0}/optout/confirm/?user='.format(get_language()), mail.outbox[1].alternatives[0][0])
+        self.assertNotIn('{0}/{1}/rsvp/'.format(settings.DOMAIN, get_language()), mail.outbox[1].alternatives[0][0])
+        self.assertIn('{0}/{1}/facilitator/'.format(settings.DOMAIN, get_language()), mail.outbox[1].alternatives[0][0])
+        self.assertNotIn('{0}/{1}/optout/confirm/?user='.format(settings.DOMAIN, get_language()), mail.outbox[1].alternatives[0][0])
 
 
     @patch('studygroups.tasks.send_message')
@@ -364,10 +357,10 @@ class TestStudyGroupTasks(TestCase):
         application.save()
         mail.outbox = []
         generate_all_meetings(sg)
-        generate_reminder(sg)
+        tasks.generate_reminder(sg)
         self.assertEqual(Reminder.objects.all().count(), 1)
         reminder = Reminder.objects.all()[0]
-        send_reminder(reminder)
+        tasks.send_reminder(reminder)
         self.assertEqual(len(mail.outbox), 2)
         #self.assertEqual(mail.outbox[0].subject, mail_data['email_subject'])
         self.assertFalse(send_message.called)
@@ -410,7 +403,7 @@ class TestStudyGroupTasks(TestCase):
 
 
         with freeze_time("2018-08-28 10:01:00"):
-            send_weekly_update()
+            tasks.send_weekly_update()
 
         self.assertEqual(len(mail.outbox), 2)
         self.assertEqual(mail.outbox[0].to[0], 'organ@team.com')
@@ -441,7 +434,7 @@ class TestStudyGroupTasks(TestCase):
         meeting.save()
 
         with freeze_time("2018-08-28 10:01:00"):
-            send_weekly_update()
+            tasks.send_weekly_update()
 
         self.assertEqual(len(mail.outbox), 1)
         self.assertEqual(mail.outbox[0].to[0], 'admin@test.com')
@@ -483,27 +476,27 @@ class TestStudyGroupTasks(TestCase):
         # send time is 1 week, 2 days, and 1 hour before the last meeting
         # freeze time to 1 hour before send time
         with freeze_time("2010-02-03 16:00:00"):
-            send_learner_surveys(sg)
+            tasks.send_learner_surveys(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # 1 hour and 10 minutes after send time
         with freeze_time("2010-02-03 18:10:00"):
-            send_learner_surveys(sg)
+            tasks.send_learner_surveys(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # 30 minutes after send_time
         with freeze_time("2010-02-03 17:30:00"):
-            send_learner_surveys(sg)
+            tasks.send_learner_surveys(sg)
             self.assertEqual(len(mail.outbox), 2)
             self.assertIn('mail1@example.net', mail.outbox[0].to + mail.outbox[1].to)
             self.assertIn('mail2@example.net', mail.outbox[0].to + mail.outbox[1].to)
             a1 = Application.objects.get(email=mail.outbox[0].to[0])
-            self.assertIn('https://example.net/en/studygroup/{}/survey/?learner={}'.format(sg.uuid, a1.uuid), mail.outbox[0].body)
+            self.assertIn('{0}/en/studygroup/{1}/survey/?learner={2}'.format(settings.DOMAIN, sg.uuid, a1.uuid), mail.outbox[0].body)
 
         mail.outbox = []
         # 2 hours after send time
         with freeze_time("2010-02-03 19:00:00"):
-            send_learner_surveys(sg)
+            tasks.send_learner_surveys(sg)
             self.assertEqual(len(mail.outbox), 0)
 
 
@@ -527,19 +520,19 @@ class TestStudyGroupTasks(TestCase):
         # send time is 1 week, 2 days, and 1 hour before the last meeting
         # freeze time to 1 day after send time
         with freeze_time("2010-01-28 17:55:34"):
-            send_facilitator_survey(sg)
+            tasks.send_facilitator_survey(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 3 days after send time
         with freeze_time("2010-01-26 17:55:34"):
-            send_facilitator_survey(sg)
+            tasks.send_facilitator_survey(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 2 days after send time
         with freeze_time("2010-01-27 17:55:34"):
-            send_facilitator_survey(sg)
+            tasks.send_facilitator_survey(sg)
             self.assertEqual(len(mail.outbox), 1)
-            self.assertIn('https://example.net/en/studygroup/{0}/facilitator_survey/'.format(sg.id), mail.outbox[0].body)
+            self.assertIn('{0}/en/studygroup/{1}/facilitator_survey/'.format(settings.DOMAIN, sg.id), mail.outbox[0].body)
             self.assertIn(sg.facilitator.email, mail.outbox[0].to)
 
 
@@ -563,19 +556,19 @@ class TestStudyGroupTasks(TestCase):
         # send time is 2 days before the last meeting
         # freeze time to 1 hour before send time
         with freeze_time("2010-02-03 17:00:00"):
-            send_facilitator_survey_reminder(sg)
+            tasks.send_facilitator_survey_reminder(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 2 hours after send time
         with freeze_time("2010-02-03 20:00:00"):
-            send_facilitator_survey_reminder(sg)
+            tasks.send_facilitator_survey_reminder(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 30 minutes after send time
         with freeze_time("2010-02-03 18:30:00"):
-            send_facilitator_survey_reminder(sg)
+            tasks.send_facilitator_survey_reminder(sg)
             self.assertEqual(len(mail.outbox), 1)
-            self.assertIn('https://example.net/en/studygroup/{0}/facilitator_survey/'.format(sg.id), mail.outbox[0].body)
+            self.assertIn('{0}/en/studygroup/{1}/facilitator_survey/'.format(settings.DOMAIN, sg.id), mail.outbox[0].body)
             self.assertIn(sg.facilitator.email, mail.outbox[0].to)
 
 
@@ -613,21 +606,21 @@ class TestStudyGroupTasks(TestCase):
         # send time is 2 days before the last meeting
         # freeze time to 1 hour before send time
         with freeze_time("2010-02-07 17:00:00"):
-            send_final_learning_circle_report(sg)
+            tasks.send_final_learning_circle_report(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 2 hours after send time
         with freeze_time("2010-02-07 20:00:00"):
-            send_final_learning_circle_report(sg)
+            tasks.send_final_learning_circle_report(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to 30 minutes after send time
         with freeze_time("2010-02-07 18:30:00"):
-            send_final_learning_circle_report(sg)
+            tasks.send_final_learning_circle_report(sg)
             self.assertIn('mail1@example.net', mail.outbox[0].to)
             self.assertIn('facilitator@example.net', mail.outbox[0].to)
             self.assertEqual(len(mail.outbox[0].to), 2)
-            self.assertIn('https://example.net/en/studygroup/{0}/report/'.format(sg.id), mail.outbox[0].body)
+            self.assertIn('{0}/en/studygroup/{1}/report/'.format(settings.DOMAIN, sg.id), mail.outbox[0].body)
             self.assertIn(sg.facilitator.email, mail.outbox[0].to)
 
 
@@ -644,19 +637,19 @@ class TestStudyGroupTasks(TestCase):
 
         # freeze time to 3 days before last meeting
         with freeze_time("2018-01-03 17:59:00"):
-            send_last_week_group_activity(sg)
+            tasks.send_last_week_group_activity(sg)
             self.assertEqual(len(mail.outbox), 0)
 
         # freeze time to less than 2 days before last meeting
         with freeze_time("2018-01-04 18:01:00"):
-            send_last_week_group_activity(sg)
+            tasks.send_last_week_group_activity(sg)
             self.assertEqual(len(mail.outbox), 1)
             self.assertIn(sg.facilitator.email, mail.outbox[0].to)
 
         mail.outbox = []
         # freeze time to less than 1 day before last meeting
         with freeze_time("2018-01-05 18:01:00"):
-            send_last_week_group_activity(sg)
+            tasks.send_last_week_group_activity(sg)
             self.assertEqual(len(mail.outbox), 0)
