+++ templates/studygroups/discourse_table.html
@@ -5,9 +5,7 @@
     <tr>
       <td>Topic</td>
       <td>Category</td>
-      <td>Replies</td>
       <td>Views</td>
-      <td>Last activity</td>
     </tr>
   </thead>
   <tbody>
@@ -15,9 +13,7 @@
       <tr>
         <td><a href="https://community.p2pu.org/t/{{ topic.slug }}">{{ topic.title }}</a></td>
         <td>{{ topic.category_id|category_title:discourse_categories }}</td>
-        <td>{{ topic.reply_count }}</td>
         <td>{{ topic.views }}</td>
-        <td>{{ topic.last_posted_at | parse_date }}</td>
       </tr>
     {% endfor %}
   </tbody>
